fix insecure openstack authentication (release) (#278)
