+++ k8s/migration/internal/controller/migrationtemplate_controller.go
@@ -23,6 +23,7 @@ import (
 
 	"github.com/go-logr/logr"
 	vjailbreakv1alpha1 "github.com/platform9/vjailbreak/k8s/migration/api/v1alpha1"
+	utils "github.com/platform9/vjailbreak/k8s/migration/pkg/utils"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -88,7 +89,7 @@ func (r *MigrationTemplateReconciler) Reconcile(ctx context.Context, req ctrl.Re
 		return ctrl.Result{}, err
 	}
 
-	openstackinfo, err := GetOpenstackInfo(ctx, openstackcreds)
+	openstackinfo, err := utils.GetOpenstackInfo(ctx, openstackcreds)
 	if err != nil {
 		r.ctxlog.Error(err, "Failed to get info of Openstack")
 		return ctrl.Result{}, err
