+++ k8s/migration/pkg/utils/credutils.go
@@ -2,13 +2,33 @@ package utils
 
 import (
 	"context"
+	"crypto/tls"
+	"crypto/x509"
+	"encoding/pem"
 	"fmt"
+	"net/http"
+	"net/url"
+	"strings"
 
+	"github.com/pkg/errors"
+	vjailbreakv1alpha1 "github.com/platform9/vjailbreak/k8s/migration/api/v1alpha1"
+	"sigs.k8s.io/controller-runtime/pkg/log"
+
+	"github.com/gophercloud/gophercloud"
+	"github.com/gophercloud/gophercloud/openstack"
+	"github.com/gophercloud/gophercloud/openstack/blockstorage/v3/volumetypes"
+	"github.com/gophercloud/gophercloud/openstack/networking/v2/networks"
+	"github.com/gophercloud/gophercloud/openstack/networking/v2/ports"
 	"github.com/platform9/vjailbreak/k8s/migration/pkg/constants"
-
 	corev1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
+
+type OpenStackClients struct {
+	BlockStorageClient *gophercloud.ServiceClient
+	ComputeClient      *gophercloud.ServiceClient
+	NetworkingClient   *gophercloud.ServiceClient
+}
 
 // VMwareCredentials holds the actual credentials after decoding
 type VMwareCredentials struct {
@@ -36,11 +56,11 @@ func GetVMwareCredentials(ctx context.Context, secretName string) (VMwareCredent
 	// Get In cluster client
 	c, err := GetInclusterClient()
 	if err != nil {
-		return VMwareCredentials{}, fmt.Errorf("failed to get in cluster client: %w", err)
+		return VMwareCredentials{}, errors.Wrap(err, "failed to get in cluster client")
 	}
 
 	if err := c.Get(ctx, client.ObjectKey{Namespace: constants.NamespaceMigrationSystem, Name: secretName}, secret); err != nil {
-		return VMwareCredentials{}, fmt.Errorf("failed to get secret '%s': %w", secretName, err)
+		return VMwareCredentials{}, errors.Wrapf(err, "failed to get secret '%s'", secretName)
 	}
 
 	if secret.Data == nil {
@@ -53,16 +73,16 @@ func GetVMwareCredentials(ctx context.Context, secretName string) (VMwareCredent
 	insecureStr := string(secret.Data["VCENTER_INSECURE"])
 
 	if host == "" {
-		return VMwareCredentials{}, fmt.Errorf("VCENTER_HOST is missing in secret '%s'", secretName)
+		return VMwareCredentials{}, errors.Errorf("VCENTER_HOST is missing in secret '%s'", secretName)
 	}
 	if username == "" {
-		return VMwareCredentials{}, fmt.Errorf("VCENTER_USERNAME is missing in secret '%s'", secretName)
+		return VMwareCredentials{}, errors.Errorf("VCENTER_USERNAME is missing in secret '%s'", secretName)
 	}
 	if password == "" {
-		return VMwareCredentials{}, fmt.Errorf("VCENTER_PASSWORD is missing in secret '%s'", secretName)
+		return VMwareCredentials{}, errors.Errorf("VCENTER_PASSWORD is missing in secret '%s'", secretName)
 	}
 
-	insecure := insecureStr == "true"
+	insecure := strings.TrimSpace(insecureStr) == "true"
 
 	return VMwareCredentials{
 		Host:     host,
@@ -72,16 +92,16 @@ func GetVMwareCredentials(ctx context.Context, secretName string) (VMwareCredent
 	}, nil
 }
 
-// getOpenStackCreds retrieves and checks the secret
+// GetOpenstackCredentials retrieves and checks the secret
 func GetOpenstackCredentials(ctx context.Context, secretName string) (OpenStackCredentials, error) {
 	secret := &corev1.Secret{}
 	// Get In cluster client
 	c, err := GetInclusterClient()
 	if err != nil {
-		return OpenStackCredentials{}, fmt.Errorf("failed to get in cluster client: %w", err)
+		return OpenStackCredentials{}, errors.Wrap(err, "failed to get in cluster client")
 	}
 	if err := c.Get(ctx, client.ObjectKey{Namespace: constants.NamespaceMigrationSystem, Name: secretName}, secret); err != nil {
-		return OpenStackCredentials{}, fmt.Errorf("failed to get secret: %w", err)
+		return OpenStackCredentials{}, errors.Wrap(err, "failed to get secret")
 	}
 
 	// Extract and validate each field
@@ -91,16 +111,17 @@ func GetOpenstackCredentials(ctx context.Context, secretName string) (OpenStackC
 		"Username":   string(secret.Data["OS_USERNAME"]),
 		"Password":   string(secret.Data["OS_PASSWORD"]),
 		"TenantName": string(secret.Data["OS_TENANT_NAME"]),
+		"RegionName": string(secret.Data["OS_REGION_NAME"]),
 	}
 
 	for key, value := range fields {
 		if value == "" {
-			return OpenStackCredentials{}, fmt.Errorf("%s is missing in secret '%s'", key, secretName)
+			return OpenStackCredentials{}, errors.Errorf("%s is missing in secret '%s'", key, secretName)
 		}
 	}
 
 	insecureStr := string(secret.Data["OS_INSECURE"])
-	insecure := insecureStr == "true"
+	insecure := strings.TrimSpace(insecureStr) == "true"
 
 	return OpenStackCredentials{
 		AuthURL:    fields["AuthURL"],
@@ -111,4 +132,259 @@ func GetOpenstackCredentials(ctx context.Context, secretName string) (OpenStackC
 		TenantName: fields["TenantName"],
 		Insecure:   insecure,
 	}, nil
+}
+
+func GetCert(endpoint string) (*x509.Certificate, error) {
+	conf := &tls.Config{
+		//nolint:gosec // This is required to skip certificate verification
+		InsecureSkipVerify: true,
+	}
+	parsedURL, err := url.Parse(endpoint)
+	if err != nil {
+		return nil, errors.Wrap(err, "error parsing URL")
+	}
+	hostname := parsedURL.Hostname()
+	conn, err := tls.Dial("tcp", hostname+":443", conf)
+	if err != nil {
+		return nil, errors.Wrapf(err, "error connecting to %s", hostname)
+	}
+	defer conn.Close()
+	cert := conn.ConnectionState().PeerCertificates[0]
+	return cert, nil
+}
+
+//nolint:dupl // This function is similar to VerifyNetworks, excluding from linting to keep it readable
+func VerifyNetworks(ctx context.Context, openstackcreds *vjailbreakv1alpha1.OpenstackCreds, targetnetworks []string) error {
+	openstackClients, err := GetOpenStackClients(ctx, openstackcreds)
+	if err != nil {
+		return errors.Wrap(err, "failed to get openstack clients")
+	}
+	allPages, err := networks.List(openstackClients.NetworkingClient, nil).AllPages()
+	if err != nil {
+		return errors.Wrap(err, "failed to list networks")
+	}
+
+	allNetworks, err := networks.ExtractNetworks(allPages)
+	if err != nil {
+		return errors.Wrap(err, "failed to extract all networks")
+	}
+
+	// Build a map of all networks
+	networkMap := make(map[string]bool)
+	for i := 0; i < len(allNetworks); i++ {
+		networkMap[allNetworks[i].Name] = true
+	}
+
+	// Verify that all network names in targetnetworks exist in the openstack networks
+	for _, targetNetwork := range targetnetworks {
+		if _, found := networkMap[targetNetwork]; !found {
+			return fmt.Errorf("network '%s' not found in OpenStack", targetNetwork)
+		}
+	}
+	return nil
+}
+
+//nolint:dupl // This function is similar to VerifyNetworks, excluding from linting to keep it readable
+func VerifyPorts(ctx context.Context, openstackcreds *vjailbreakv1alpha1.OpenstackCreds, targetports []string) error {
+	openstackClients, err := GetOpenStackClients(ctx, openstackcreds)
+	if err != nil {
+		return errors.Wrap(err, "failed to get openstack clients")
+	}
+
+	allPages, err := ports.List(openstackClients.NetworkingClient, nil).AllPages()
+	if err != nil {
+		return errors.Wrap(err, "failed to list ports")
+	}
+
+	allPorts, err := ports.ExtractPorts(allPages)
+	if err != nil {
+		return errors.Wrap(err, "failed to extract all ports")
+	}
+
+	// Build a map of all ports
+	portMap := make(map[string]bool)
+	for i := 0; i < len(allPorts); i++ {
+		portMap[allPorts[i].ID] = true
+	}
+
+	// Verify that all port names in targetports exist in the openstack ports
+	for _, targetPort := range targetports {
+		if _, found := portMap[targetPort]; !found {
+			return errors.Wrap(fmt.Errorf("port '%s' not found in OpenStack", targetPort), "failed to verify ports")
+		}
+	}
+	return nil
+}
+
+func VerifyStorage(ctx context.Context, openstackcreds *vjailbreakv1alpha1.OpenstackCreds, targetstorages []string) error {
+	openstackClients, err := GetOpenStackClients(ctx, openstackcreds)
+	if err != nil {
+		return err
+	}
+	allPages, err := volumetypes.List(openstackClients.BlockStorageClient, nil).AllPages()
+	if err != nil {
+		return fmt.Errorf("failed to list volume types: %w", err)
+	}
+
+	allvoltypes, err := volumetypes.ExtractVolumeTypes(allPages)
+	if err != nil {
+		return fmt.Errorf("failed to extract all volume types: %w", err)
+	}
+
+	// Verify that all volume types in targetstorage exist in the openstack volume types
+	for _, targetstorage := range targetstorages {
+		found := false
+		for i := 0; i < len(allvoltypes); i++ {
+			if allvoltypes[i].Name == targetstorage {
+				found = true
+				break
+			}
+		}
+		if !found {
+			return fmt.Errorf("volume type '%s' not found in OpenStack", targetstorage)
+		}
+	}
+	return nil
+}
+
+func GetOpenstackInfo(ctx context.Context, openstackcreds *vjailbreakv1alpha1.OpenstackCreds) (*vjailbreakv1alpha1.OpenstackInfo, error) {
+	var openstackvoltypes []string
+	var openstacknetworks []string
+	openstackClients, err := GetOpenStackClients(ctx, openstackcreds)
+	if err != nil {
+		return nil, err
+	}
+	allVolumeTypePages, err := volumetypes.List(openstackClients.BlockStorageClient, nil).AllPages()
+	if err != nil {
+		return nil, fmt.Errorf("failed to list volume types: %w", err)
+	}
+
+	allvoltypes, err := volumetypes.ExtractVolumeTypes(allVolumeTypePages)
+	if err != nil {
+		return nil, fmt.Errorf("failed to extract all volume types: %w", err)
+	}
+
+	for i := 0; i < len(allvoltypes); i++ {
+		openstackvoltypes = append(openstackvoltypes, allvoltypes[i].Name)
+	}
+
+	allNetworkPages, err := networks.List(openstackClients.NetworkingClient, nil).AllPages()
+	if err != nil {
+		return nil, fmt.Errorf("failed to list networks: %w", err)
+	}
+
+	allNetworks, err := networks.ExtractNetworks(allNetworkPages)
+	if err != nil {
+		return nil, fmt.Errorf("failed to extract all networks: %w", err)
+	}
+
+	for i := 0; i < len(allNetworks); i++ {
+		openstacknetworks = append(openstacknetworks, allNetworks[i].Name)
+	}
+
+	return &vjailbreakv1alpha1.OpenstackInfo{
+		VolumeTypes: openstackvoltypes,
+		Networks:    openstacknetworks,
+	}, nil
+}
+
+// GetOpenStackClients is a function to create openstack clients
+func GetOpenStackClients(ctx context.Context, openstackcreds *vjailbreakv1alpha1.OpenstackCreds) (*OpenStackClients, error) {
+	if openstackcreds == nil {
+		return nil, fmt.Errorf("openstackcreds cannot be nil")
+	}
+
+	openstackCredential, err := GetOpenstackCredentials(ctx, openstackcreds.Spec.SecretRef.Name)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to get openstack credentials from secret")
+	}
+
+	endpoint := gophercloud.EndpointOpts{
+		Region: openstackCredential.RegionName,
+	}
+	providerClient, err := ValidateAndGetProviderClient(ctx, openstackcreds)
+	if err != nil {
+		return nil, errors.Wrap(err, fmt.Sprintf("failed to get provider client for region '%s'", openstackCredential.RegionName))
+	}
+	if providerClient == nil {
+		return nil, fmt.Errorf("failed to get provider client for region '%s'", openstackCredential.RegionName)
+	}
+	computeClient, err := openstack.NewComputeV2(providerClient, endpoint)
+	if err != nil {
+		return nil, errors.Wrap(err, fmt.Sprintf("failed to create openstack compute client for region '%s'", openstackCredential.RegionName))
+	}
+	blockStorageClient, err := openstack.NewBlockStorageV3(providerClient, endpoint)
+	if err != nil {
+		return nil, errors.Wrap(err, fmt.Sprintf("failed to create openstack block storage client for region '%s'",
+			openstackCredential.RegionName))
+	}
+	networkingClient, err := openstack.NewNetworkV2(providerClient, endpoint)
+	if err != nil {
+		return nil, errors.Wrap(err, fmt.Sprintf("failed to create openstack networking client for region '%s'",
+			openstackCredential.RegionName))
+	}
+	return &OpenStackClients{
+		BlockStorageClient: blockStorageClient,
+		ComputeClient:      computeClient,
+		NetworkingClient:   networkingClient,
+	}, nil
+}
+
+// ValidateAndGetProviderClient is a function to get provider client
+func ValidateAndGetProviderClient(ctx context.Context,
+	openstackcreds *vjailbreakv1alpha1.OpenstackCreds) (*gophercloud.ProviderClient, error) {
+	ctxlog := log.FromContext(ctx)
+	openstackCredential, err := GetOpenstackCredentials(ctx, openstackcreds.Spec.SecretRef.Name)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to get openstack credentials from secret")
+	}
+
+	providerClient, err := openstack.NewClient(openstackCredential.AuthURL)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to create openstack client")
+	}
+	tlsConfig := &tls.Config{
+		MinVersion: tls.VersionTLS12,
+	}
+	if openstackCredential.Insecure {
+		ctxlog.Info("Insecure flag is set, skipping certificate verification")
+		tlsConfig.InsecureSkipVerify = true
+	} else {
+		// Get the certificate for the Openstack endpoint
+		caCert, certerr := GetCert(openstackCredential.AuthURL)
+		if certerr != nil {
+			return nil, errors.Wrap(certerr, "failed to get certificate for openstack")
+		}
+		// Logging the certificate
+		ctxlog.Info(fmt.Sprintf("Trusting certificate for '%s'", openstackCredential.AuthURL))
+		ctxlog.Info(string(pem.EncodeToMemory(&pem.Block{
+			Type:  "CERTIFICATE",
+			Bytes: caCert.Raw,
+		})))
+		// Trying to fetch the system cert pool and add the Openstack certificate to it
+		caCertPool, _ := x509.SystemCertPool()
+		if caCertPool == nil {
+			caCertPool = x509.NewCertPool()
+		}
+		caCertPool.AddCert(caCert)
+		tlsConfig.RootCAs = caCertPool
+	}
+	transport := &http.Transport{
+		TLSClientConfig: tlsConfig,
+	}
+	providerClient.HTTPClient = http.Client{
+		Transport: transport,
+	}
+	err = openstack.Authenticate(providerClient, gophercloud.AuthOptions{
+		IdentityEndpoint: openstackCredential.AuthURL,
+		Username:         openstackCredential.Username,
+		Password:         openstackCredential.Password,
+		DomainName:       openstackCredential.DomainName,
+		TenantName:       openstackCredential.TenantName,
+	})
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to authenticate to openstack")
+	}
+
+	return providerClient, nil
 }
